Termino clases para GeoddsService. Los Test de GeoddsTest pasan correctamente. Falta agregar los metodos de municipio, localidades y distancia

+++ src/main/java/dds/grupo3/clases/miembro/Miembro.java
@@ -41,7 +41,7 @@ public class Miembro {
 	public void agregarUnTrayecto (Trayecto unTrayecto){ trayectos.add(unTrayecto); }
 
 	public float calcularHU(){
-		return trayectos.stream().mapToInt(t -> t.calcularHU()).sum();
+		return trayectos.stream().mapToInt(t -> (int) t.calcularHU()).sum();
 	}
 /*
 	public List<int> calcularHUPorcentual(){
